replace logic from repository to service

+++ src/main/java/com/example/trelloimplementation/privatedb/entity/Container.java
@@ -1,11 +1,14 @@
 package com.example.trelloimplementation.privatedb.entity;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.*;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import javax.persistence.Column;
+import javax.persistence.OneToMany;
+import javax.persistence.CascadeType;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -24,13 +27,15 @@ public class Container {
     @Column(name = "container_id")
     private UUID containerId;
 
+//todo    @NonNull
     @Column(name = "container_name")
     private String containerName;
 
+//todo    @NonNull
+//todo    @UniqueElements
     @Column(name = "column_order")
-    private Long containerOrder;
+    private int containerOrder;
 
     @OneToMany(mappedBy = "container", cascade = CascadeType.ALL)
-//    @JsonManagedReference
     private List<Task> tasks = new ArrayList<>();
 }
