+++ src/main/java/com/example/trelloimplementation/privatedb/vo/ContainerVO.java
@@ -17,7 +17,7 @@ public class ContainerVO {
 
     private String containerName;
 
-    private Long containerOrder;
+    private int containerOrder;
 
     private List<Task> tasks = new ArrayList<>();
